Various logical improvements
